Update WooCommerce hook in analytics sync
The 'woocommerce_update_order' action is now triggered after saving the order and passes both the order ID and order object. This ensures analytics are updated with the latest order data. This also ensures that the action is triggered the same way as WooCommerce implemented it, and ensures compliance with everyone else using this hook.